requirements.txt
groq
python-dotenv
speechrecognition
pydub
pyaudio
pynput
keyboard
pyautogui
pyperclip
ipykernel
gtts
elevenlabs
gradio